backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
rq
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib
python-multipart
python-dotenv
google-generativeai
